Refactoring line plot to use datum
