+++ js/src/main/scala/com/cibo/evilplot/EvilPlot.scala
@@ -73,6 +73,22 @@ object EvilPlot {
   @JSExport
   def renderExample(canvasId: String): Unit = {
     addExample(DemoPlots.simpleScatterPlot)
+    addExample(DemoPlots.densityPlot)
+    addExample(DemoPlots.legendFeatures)
+    addExample(DemoPlots.axesTesting)
+    addExample(DemoPlots.functionPlot)
+    addExample(DemoPlots.markerPlot)
+    addExample(DemoPlots.marginalHistogram)
+    addExample(DemoPlots.scatterPlot)
+    addExample(DemoPlots.barChart)
+    addExample(DemoPlots.boxPlot)
+    addExample(DemoPlots.clusteredBoxPlot)
+    addExample(DemoPlots.facetedPlot)
+    addExample(DemoPlots.heatmap)
+    addExample(DemoPlots.marginalHistogram)
+    addExample(DemoPlots.clusteredBarChart)
+    addExample(DemoPlots.stackedBarChart)
+    addExample(DemoPlots.clusteredStackedBarChart)
   }
 
   private def addExample(plot: Drawable): Unit = {
