+++ shared/src/main/scala/com/cibo/evilplot/plot/TransformUtils.scala
@@ -2,10 +2,9 @@ package com.cibo.evilplot.plot
 
 import com.cibo.evilplot.colors.Color
 import com.cibo.evilplot.geometry.{Clipping, Drawable, EmptyDrawable, Extent, LineDash, LineStyle, Path, StrokeStyle}
-import com.cibo.evilplot.numeric._
+import com.cibo.evilplot.numeric.{Datum2d, _}
 import com.cibo.evilplot.plot.BoxPlot.makePlot
 import com.cibo.evilplot.plot.LinePlot.LinePlotRenderer
-import com.cibo.evilplot.plot.PlotUtils.{PlotContext}
 import com.cibo.evilplot.plot.ScatterPlot.ScatterPlotRenderer
 import com.cibo.evilplot.plot.aesthetics.Theme
 import com.cibo.evilplot.plot.renderers.BoxRenderer.BoxRendererContext
@@ -51,6 +50,24 @@ trait TransformWorldToScreen {
 
 object TransformWorldToScreen extends TransformWorldToScreen
 
+case class PlotContext(plot: Plot,
+                       plotExtent: Extent){
+
+  lazy val xBounds: Bounds = plot.xbounds
+  lazy val yBounds: Bounds = plot.ybounds
+
+  def xCartesianTransform: Double => Double = TransformWorldToScreen.xCartesianTransformer(xBounds, plotExtent)
+  def yCartesianTransform: Double => Double = TransformWorldToScreen.yCartesianTransformer(yBounds, plotExtent)
+
+  def transformDatumToWorld[X <: Datum2d[X]](point: X): X = TransformWorldToScreen.transformDatumToWorld(point, xCartesianTransform, yCartesianTransform)
+  def transformDatumsToWorld[X <: Datum2d[X]](points: Seq[X]): Seq[X] = points.map(transformDatumToWorld)
+
+}
+
+object PlotContext {
+  def fromPlotExtent(plot: Plot, extent: Extent): PlotContext = apply(plot, extent)
+}
+
 object PlotUtils {
 
 
@@ -79,142 +96,6 @@ object PlotUtils {
     )
     (xbounds, ybounds)
   }
-
-  case class PlotContext(plot: Plot,
-                         plotExtent: Extent){
-
-    lazy val xBounds: Bounds = plot.xbounds
-    lazy val yBounds: Bounds = plot.ybounds
-
-    def xCartesianTransform: Double => Double = TransformWorldToScreen.xCartesianTransformer(xBounds, plotExtent)
-    def yCartesianTransform: Double => Double = TransformWorldToScreen.yCartesianTransformer(yBounds, plotExtent)
-  }
-
-  object PlotContext {
-    def fromPlotExtent(plot: Plot, extent: Extent): PlotContext = apply(plot, extent)
-  }
-
-  case class CartesianDataRenderer[X <: Datum2d[X]](data: Seq[X]) extends TransformWorldToScreen {
-
-    def transformWTS(plotContext: PlotContext): Seq[X] = {
-      val transformed = transformDatumToPlotSpace(
-        data,
-        plotContext.xCartesianTransform,
-        plotContext.yCartesianTransform)
-      transformed
-    }
-
-    def scatter(pointToDrawable: X => Drawable,
-                legendCtx: LegendContext = LegendContext.empty)(pCtx: PlotContext)(implicit theme: Theme): PlotRenderer = {
-      ScatterPlotRenderer(data, new PointRenderer[X] {
-        def render(index: X): Drawable = pointToDrawable(index)
-      })
-    }
-
-    def scatter(pointRenderer: PointRenderer[X])(pCtx: PlotContext)(implicit theme: Theme): ScatterPlotRenderer[X] = {
-      ScatterPlotRenderer(data, pointRenderer)
-    }
-
-//      new PlotRenderer {
-//
-//      override def legendContext: LegendContext = legendCtx
-//
-//      def render(plot: Plot, plotExtent: Extent)(implicit theme: Theme): Drawable = {
-//        val points = transformWTS(pCtx).zipWithIndex
-//          .withFilter(p => pCtx.plotExtent.contains(p._1))
-//          .flatMap {
-//            case (point, index) =>
-//              val r = pointToDrawable(point)
-//              if (r.isEmpty) None else Some(r.translate(x = point.x, y = point.y))
-//          }
-//          .group
-//
-//        points
-//      }
-//
-//    }
-
-    def manipulate(x: Seq[X] => Seq[X]): Seq[X] = x(data)
-
-    def filter(x: X => Boolean): CartesianDataRenderer[X] = this.copy(data.filter(x))
-
-    def line(
-            strokeWidth: Option[Double] = None,
-            color: Option[Color] = None,
-            label: Drawable = EmptyDrawable(),
-            lineStyle: Option[LineStyle] = None,
-            legendCtx: LegendContext = LegendContext.empty
-           )(pCtx: PlotContext)(implicit theme: Theme): PlotRenderer = {
-        LinePlotRenderer(data, PathRenderer.default(strokeWidth, color, label, lineStyle))
-    }
-
-    def line(pathRenderer: PathRenderer)(pCtx: PlotContext)(implicit theme: Theme): PlotRenderer = {
-      LinePlotRenderer(data, pathRenderer)
-    }
-
-    def boxAndWhisker(
-      dataGroupFn: Seq[X] => Seq[Seq[Double]],
-      quantiles: (Double, Double, Double) = (0.25, 0.50, 0.75),
-      spacing: Option[Double] = None,
-      boundBuffer: Option[Double] = None,
-      boxRenderer: Option[BoxRenderer] = None,
-      pointRenderer: Option[PointRenderer[BoxPlotPoint]] = None
-      )(pCtx: PlotContext)(implicit theme: Theme): Drawable = {
-        val groupedData = dataGroupFn(data)
-        val boxContexts = dataGroupFn(data).zipWithIndex.map {
-          case (dist, index) =>
-            if (dist.nonEmpty) {
-              val summary = BoxPlotSummaryStatistics(dist, quantiles)
-              Some(BoxRendererContext(summary, index))
-            } else None
-        }
-        makePlot(
-          dataGroupFn(data),
-          boxContexts,
-          spacing,
-          None,
-          boundBuffer,
-          boxRenderer,
-          pointRenderer
-        ).render(pCtx.plotExtent)
-      }
-
-    private def makePlot(
-                          data: Seq[Seq[Double]],
-                          boxContexts: Seq[Option[BoxRendererContext]],
-                          spacing: Option[Double] = None,
-                          clusterSpacing: Option[Double] = None,
-                          boundBuffer: Option[Double] = None,
-                          boxRenderer: Option[BoxRenderer] = None,
-                          pointRenderer: Option[PointRenderer[BoxPlotPoint]] = None
-                        )(implicit theme: Theme): Plot = {
-      val xbounds = Bounds(0, boxContexts.size)
-      val ybounds = Plot.expandBounds(
-        Bounds(
-          data.flatten.reduceOption[Double](math.min).getOrElse(0),
-          data.flatten.reduceOption[Double](math.max).getOrElse(0)
-        ),
-        boundBuffer.getOrElse(theme.elements.boundBuffer)
-      )
-      Plot(
-        xbounds,
-        ybounds,
-        BoxPlotRenderer(
-          boxContexts,
-          boxRenderer.getOrElse(BoxRenderer.default()),
-          pointRenderer.getOrElse(PointRenderer.default[BoxPlotPoint]()),
-          spacing.getOrElse(theme.elements.boxSpacing),
-          clusterSpacing
-        )
-      )
-    }
-
-  }
-
-
-
-
-
 
 }
