--- shared/src/main/scala/com/cibo/evilplot/plot/XyPlot.scala
@@ -1,82 +1,0 @@
-/*
- * Copyright (c) 2018, CiBO Technologies, Inc.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors
- *    may be used to endorse or promote products derived from this software without
- *    specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package com.cibo.evilplot.plot
-
-import com.cibo.evilplot.colors.{Color, RGB}
-import com.cibo.evilplot.geometry._
-import com.cibo.evilplot.numeric.{Bounds, Datum2d, Point, Point2d}
-import com.cibo.evilplot.plot.Plot.Transformer
-import com.cibo.evilplot.plot.PlotUtils.{CartesianDataRenderer, PlotContext}
-import com.cibo.evilplot.plot.TransformWorldToScreen.{createTransformers, transformDatumToPlotSpace}
-import com.cibo.evilplot.plot.aesthetics.Theme
-import com.cibo.evilplot.plot.renderers.{PathRenderer, PlotRenderer, PointRenderer}
-
-object CartesianPlot {
-
-  import TransformWorldToScreen._
-  final case class CartesianPlotRenderer(drawablesToPlot: Seq[PlotContext => PlotRenderer],
-                                         xBounds: Bounds,
-                                         yBounds: Bounds) extends PlotRenderer {
-    override def legendContext: LegendContext = LegendContext()
-
-
-    def render(plot: Plot, plotExtent: Extent)(implicit theme: Theme): Drawable = {
-      drawablesToPlot.foldLeft(EmptyDrawable() : Drawable){ case (accum, dr) =>
-        dr(PlotContext(plot, plotExtent)).render(plot, plotExtent) behind accum
-      }
-    }
-  }
-
-  type ContextToDrawable[X <: Datum2d[X]] = CartesianDataRenderer[X] => PlotContext => PlotRenderer
-
-  def apply[X <: Datum2d[X]](
-                              data: Seq[X],
-                              xboundBuffer: Option[Double] = None,
-                              yboundBuffer: Option[Double] = None
-                            )(
-          contextToDrawable: ContextToDrawable[X]*,
-  )(implicit theme: Theme): Plot = {
-
-    val (xbounds, ybounds) = PlotUtils.bounds(data, theme.elements.boundBuffer, xboundBuffer, yboundBuffer)
-
-    val cartesianDataRenderer = CartesianDataRenderer(data)
-
-    Plot(
-      xbounds,
-      ybounds,
-      CartesianPlotRenderer(
-        contextToDrawable.map(x => x(cartesianDataRenderer)),
-        xbounds,
-        ybounds
-      )
-    )
-  }
-}
